requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
python-magic==0.4.27
pytest-randomly==3.16.0
pytest-xdist==3.6.1
pypdf==5.1.0
pytesseract==0.3.13
pdfplumber==0.11.4
PyMuPDF==1.24.13
scikit-learn ==1.5.2
pandas==2.2.3
joblib==1.4.2
serpapi==0.1.5
pillow==11.0.0
requests==2.32.3
beautifulsoup4==4.12.3
faker==33.0.0
python-docx 
openpyxl==3.1.5
pillow==11.0.0
pymupdf==1.24.13